services/ingestion/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
pdfminer.six==20231228
python-docx==1.1.0
minio==7.2.5
python-multipart==0.0.9
pydantic==2.7.0
chardet==5.2.0
unicodedata2==15.1.0
